add menu & merge someone's code

+++ source/binary-insertion-sort.cpp
@@ -16,7 +16,7 @@ void binaryInsertionSort(vector<int>& nums){
         nums[left] = key;
     }
 }
-void binaryInsertionSortOperationsCount(vector<int>& nums, int& assignments, int& comparisons){
+void binaryInsertionSortOperationsCount(vector<int>& nums, long long& assignments, long long& comparisons){
     assignments = comparisons = 0;
     ++assignments; // before loop
     for (int i = 1; ++comparisons && i < nums.size(); i++, ++assignments){
